Add-Chatting
Add the ability to send messages to the chat through the overlay, sometimes these messages take abit to appear, as they need to be received on the WebSocket before they appear. I couldnt figure out a way to append them the correct way.

The messages are sent directly via the web socket, I've added a function that hooks into the existing web socket connections and allows us to intercept sent/received messages.
as well as send our own messages.

the functionality of sending the messages relies on the @msgpack/msgpack libary which will need to be installed. I didn't add the changes to package.json here but let me know if you want me to. This library is needed in order to encode the WebSocket messages as MessagePack.

Ive also done some styling changes to the overlay, so its abit clearer and doesn't interfere with the stream as much.